Add files via upload
Using dendrogram to find the optimum number of clusters and then Using k means algorithm to cluster people based on common inputs, I have created 3 clusters to map people. I have also created a plot to depict if gender played an import part in increasing the odds of survival and the stats of women vs men survival in the disaster.